+++ opentelemetry-sdk/tests/metrics/test_aggregation.py
@@ -624,12 +624,14 @@ class TestHistogramConvertAggregationTemporality(TestCase):
     def test_previous_point_none(self):
 
         current_point = HistogramPoint(
-            start_time_unix_nano=0,
-            time_unix_nano=1,
+            aggregation_temporality=AggregationTemporality.DELTA,
             bucket_counts=[0, 2, 1, 2, 0],
             explicit_bounds=[0, 5, 10, 25],
+            max=24,
+            min=2,
+            start_time_unix_nano=0,
             sum=70,
-            aggregation_temporality=AggregationTemporality.DELTA,
+            time_unix_nano=1,
         )
 
         self.assertEqual(
@@ -648,42 +650,50 @@ class TestHistogramConvertAggregationTemporality(TestCase):
 
             _convert_aggregation_temporality(
                 HistogramPoint(
-                    start_time_unix_nano=0,
-                    time_unix_nano=1,
+                    aggregation_temporality=AggregationTemporality.DELTA,
                     bucket_counts=[0, 2, 1, 2, 0],
                     explicit_bounds=[0, 5, 10, 25],
+                    max=24,
+                    min=2,
+                    start_time_unix_nano=0,
                     sum=70,
-                    aggregation_temporality=AggregationTemporality.DELTA,
+                    time_unix_nano=1,
                 ),
                 HistogramPoint(
-                    start_time_unix_nano=1,
-                    time_unix_nano=2,
+                    aggregation_temporality=AggregationTemporality.DELTA,
                     bucket_counts=[0, 1, 3, 0, 0],
                     explicit_bounds=[0, 5, 10, 25],
+                    max=9,
+                    min=2,
+                    start_time_unix_nano=1,
                     sum=35,
-                    aggregation_temporality=AggregationTemporality.DELTA,
+                    time_unix_nano=2,
                 ),
                 AggregationTemporality.DELTA,
             ),
 
     def test_same_aggregation_temporality_cumulative(self):
         current_point = HistogramPoint(
-            start_time_unix_nano=0,
-            time_unix_nano=2,
+            aggregation_temporality=AggregationTemporality.CUMULATIVE,
             bucket_counts=[0, 3, 4, 2, 0],
             explicit_bounds=[0, 5, 10, 25],
+            max=24,
+            min=1,
+            start_time_unix_nano=0,
             sum=105,
-            aggregation_temporality=AggregationTemporality.CUMULATIVE,
+            time_unix_nano=2,
         )
         self.assertEqual(
             _convert_aggregation_temporality(
                 HistogramPoint(
-                    start_time_unix_nano=0,
-                    time_unix_nano=1,
+                    aggregation_temporality=AggregationTemporality.CUMULATIVE,
                     bucket_counts=[0, 2, 1, 2, 0],
                     explicit_bounds=[0, 5, 10, 25],
+                    max=24,
+                    min=1,
+                    start_time_unix_nano=0,
                     sum=70,
-                    aggregation_temporality=AggregationTemporality.CUMULATIVE,
+                    time_unix_nano=1,
                 ),
                 current_point,
                 AggregationTemporality.CUMULATIVE,
@@ -693,23 +703,27 @@ class TestHistogramConvertAggregationTemporality(TestCase):
 
     def test_same_aggregation_temporality_delta(self):
         current_point = HistogramPoint(
-            start_time_unix_nano=1,
-            time_unix_nano=2,
+            aggregation_temporality=AggregationTemporality.DELTA,
             bucket_counts=[0, 1, 3, 0, 0],
             explicit_bounds=[0, 5, 10, 25],
+            max=9,
+            min=1,
+            start_time_unix_nano=1,
             sum=35,
-            aggregation_temporality=AggregationTemporality.DELTA,
+            time_unix_nano=2,
         )
 
         self.assertEqual(
             _convert_aggregation_temporality(
                 HistogramPoint(
-                    start_time_unix_nano=0,
-                    time_unix_nano=2,
+                    aggregation_temporality=AggregationTemporality.CUMULATIVE,
                     bucket_counts=[0, 3, 4, 2, 0],
                     explicit_bounds=[0, 5, 10, 25],
+                    max=24,
+                    min=1,
+                    start_time_unix_nano=0,
                     sum=105,
-                    aggregation_temporality=AggregationTemporality.CUMULATIVE,
+                    time_unix_nano=2,
                 ),
                 current_point,
                 AggregationTemporality.DELTA,
@@ -719,67 +733,79 @@ class TestHistogramConvertAggregationTemporality(TestCase):
 
     def test_aggregation_temporality_to_cumulative(self):
         current_point = HistogramPoint(
-            start_time_unix_nano=1,
-            time_unix_nano=2,
+            aggregation_temporality=AggregationTemporality.DELTA,
             bucket_counts=[0, 1, 3, 0, 0],
             explicit_bounds=[0, 5, 10, 25],
+            max=24,
+            min=1,
+            start_time_unix_nano=1,
             sum=35,
-            aggregation_temporality=AggregationTemporality.DELTA,
+            time_unix_nano=2,
         )
 
         self.assertEqual(
             _convert_aggregation_temporality(
                 HistogramPoint(
-                    start_time_unix_nano=0,
-                    time_unix_nano=1,
+                    aggregation_temporality=AggregationTemporality.CUMULATIVE,
                     bucket_counts=[0, 2, 1, 2, 0],
                     explicit_bounds=[0, 5, 10, 25],
+                    max=24,
+                    min=1,
+                    start_time_unix_nano=0,
                     sum=70,
-                    aggregation_temporality=AggregationTemporality.CUMULATIVE,
+                    time_unix_nano=1,
                 ),
                 current_point,
                 AggregationTemporality.CUMULATIVE,
             ),
             HistogramPoint(
-                start_time_unix_nano=0,
-                time_unix_nano=2,
+                aggregation_temporality=AggregationTemporality.CUMULATIVE,
                 bucket_counts=[0, 3, 4, 2, 0],
                 explicit_bounds=[0, 5, 10, 25],
+                max=24,
+                min=1,
+                start_time_unix_nano=0,
                 sum=105,
-                aggregation_temporality=AggregationTemporality.CUMULATIVE,
+                time_unix_nano=2,
             ),
         )
 
     def test_aggregation_temporality_to_delta(self):
         current_point = HistogramPoint(
-            start_time_unix_nano=0,
-            time_unix_nano=2,
+            aggregation_temporality=AggregationTemporality.CUMULATIVE,
             bucket_counts=[0, 3, 4, 2, 0],
             explicit_bounds=[0, 5, 10, 25],
+            max=22,
+            min=3,
+            start_time_unix_nano=0,
             sum=105,
-            aggregation_temporality=AggregationTemporality.CUMULATIVE,
+            time_unix_nano=2,
         )
 
         self.assertEqual(
             _convert_aggregation_temporality(
                 HistogramPoint(
-                    start_time_unix_nano=0,
-                    time_unix_nano=1,
+                    aggregation_temporality=AggregationTemporality.CUMULATIVE,
                     bucket_counts=[0, 2, 1, 2, 0],
                     explicit_bounds=[0, 5, 10, 25],
+                    max=24,
+                    min=1,
+                    start_time_unix_nano=0,
                     sum=70,
-                    aggregation_temporality=AggregationTemporality.CUMULATIVE,
+                    time_unix_nano=1,
                 ),
                 current_point,
                 AggregationTemporality.DELTA,
             ),
             HistogramPoint(
-                start_time_unix_nano=1,
-                time_unix_nano=2,
+                aggregation_temporality=AggregationTemporality.DELTA,
                 bucket_counts=[0, 1, 3, 0, 0],
                 explicit_bounds=[0, 5, 10, 25],
+                max=22,
+                min=3,
+                start_time_unix_nano=1,
                 sum=35,
-                aggregation_temporality=AggregationTemporality.DELTA,
+                time_unix_nano=2,
             ),
         )
 
@@ -871,7 +897,7 @@ class TestDefaultAggregation(TestCase):
     def test_observable_counter(self):
 
         aggregation = self.default_aggregation._create_aggregation(
-            ObservableCounter(Mock(), Mock(), Mock(), Mock())
+            ObservableCounter(Mock(), Mock(), Mock(), callbacks=[Mock()])
         )
         self.assertIsInstance(aggregation, _SumAggregation)
         self.assertTrue(aggregation._instrument_is_monotonic)
@@ -883,7 +909,7 @@ class TestDefaultAggregation(TestCase):
     def test_observable_up_down_counter(self):
 
         aggregation = self.default_aggregation._create_aggregation(
-            ObservableUpDownCounter(Mock(), Mock(), Mock(), Mock())
+            ObservableUpDownCounter(Mock(), Mock(), Mock(), callbacks=[Mock()])
         )
         self.assertIsInstance(aggregation, _SumAggregation)
         self.assertFalse(aggregation._instrument_is_monotonic)
@@ -912,7 +938,7 @@ class TestDefaultAggregation(TestCase):
                 Mock(),
                 Mock(),
                 Mock(),
-                Mock(),
+                callbacks=[Mock()],
             )
         )
         self.assertIsInstance(aggregation, _LastValueAggregation)