+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/open-telemetry/opentelemetry-python/compare/v1.10.0-0.29b0...HEAD)
 
+- Add support for zero or more callbacks
+  ([#2602](https://github.com/open-telemetry/opentelemetry-python/pull/2602))
 - Fix parsing of trace flags when extracting traceparent
   ([#2577](https://github.com/open-telemetry/opentelemetry-python/pull/2577))
 - Add default aggregation
@@ -30,6 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#2547](https://github.com/open-telemetry/opentelemetry-python/pull/2547))
 - Update otlp-proto-grpc and otlp-proto-http exporters to have more lax requirements for `backoff` lib
   ([#2575](https://github.com/open-telemetry/opentelemetry-python/pull/2575))
+- Add min/max to histogram point
+  ([#2581](https://github.com/open-telemetry/opentelemetry-python/pull/2581))
 
 ## [1.10.0-0.29b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.10.0-0.29b0) - 2022-03-10